third_party/WebKit/LayoutTests/platform/mac/fast/events/context-nodrag-expected.txt

This tests to make sure that right clicking when the left-mouse button is pressed disables the drag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL window.getSelection().isCollapsed should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
This is a test box. ABC 123
